mysite/requirements-dev.txt
asgiref==3.9.2
Django==5.2.6
factory_boy==3.3.3
Faker==37.8.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
sqlparse==0.5.3
tzdata==2025.2
